+++ podpointclient/endpoints.py
@@ -7,6 +7,7 @@ PODS = '/pods'
 UNITS = '/units'
 CHARGE_SCHEDULES = '/charge-schedules'
 CHARGES = '/charges'
+FIRMWARE = '/firmware'
 
 API_BASE = 'api.pod-point.com'
 API_VERSION = 'v4'